backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# Redis和任务队列
redis==5.0.1
celery==5.3.4
flower==2.0.1

# AI模型
openai==1.3.5
anthropic==0.7.1
tiktoken==0.5.1

# 代码分析
tree-sitter==0.20.4
tree-sitter-languages==1.8.0

# Git操作
GitPython==3.1.40

# 测试和覆盖率
pytest==7.4.3
pytest-asyncio==0.21.1
coverage==7.3.2

# 工具
python-dotenv==1.0.0
httpx==0.25.2
websockets==12.0
aiofiles==23.2.1
python-multipart==0.0.6

# 日志
loguru==0.7.2

# 安全
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# 数据处理
pandas==2.1.3
numpy==1.26.2

# HTTP客户端
requests==2.31.0

